dump: don't overwrite existing backup directory
Also move our memFS implementation into the `testutil` package so it can
be used by other packages as well.

closes: https://github.com/planetscale/project-big-bang/issues/311
